move configs to config package
